refactor: move Z-Wave specific logic from coordinator to provider
- Move sensor-based event detection (_handle_lock_state_change) to
  ZWaveJSLockProvider's subscribe_lock_events method
- Move LOCK_ACTIVITY_MAP and LOCK_STATE_MAP to providers/zwave_js.py
- Create providers/const.py for provider-specific constants
- Make get_activity_for_sensor_event opt-in (not in base class)
- Add _update_listeners call after provider connects to fix event
  subscription timing for restored locks
- Fix refresh condition to only query lock when slot has code but
  value is unknown (addresses Copilot review feedback)
- Remove unnecessary event_label initialization
- Update PROVIDERS.md to match actual PROVIDER_MAP pattern
