+++ custom_components/keymaster/providers/const.py
@@ -1,0 +1,8 @@
+"""Constants for lock providers."""
+
+# Z-Wave sensor types for lock events
+ALARM_TYPE = "alarm_type"
+ACCESS_CONTROL = "access_control"
+
+# Sentinel value for unknown/unsupported values in activity maps
+UNKNOWN = "unknown"
